+++ src/app/components/admin/notices/NoticeGet.tsx
@@ -1,18 +1,14 @@
 'use client';
 
-import {
-  ALERT_TEXT,
-  BTN_TEXT,
-  INPUT_TEXT,
-  TABLE_TEXT,
-} from '@/app/constants/admin';
+import { BTN_TEXT, INPUT_TEXT, TABLE_TEXT } from '@/app/constants/admin';
 import { useCallback, useEffect, useState } from 'react';
-import { callDelete, callGet, callPut } from '@/app/utils/callApi';
+import { callGet } from '@/app/utils/callApi';
 import Button from '../../common/Button';
 import Input from '../../common/Input';
 import NoticePost from './NoticePost';
 import Icons from '../../common/Icons';
 import { ArrowDownIcon } from '@/app/ui/iconPath';
+import NoticesPut from './NoticePut';
 
 export default function NoticeGet() {
   const [inputId, setInputId] = useState('');
@@ -68,45 +64,6 @@ export default function NoticeGet() {
     setEditingId(null); // 새로운 항목을 열 때 편집 모드 초기화
   };
 
-  // 공지사항 수정모드
-  const handleEdit = (notice_id: number, title: string, content: string) => {
-    setEditingId(notice_id);
-    setEditNotice({
-      title,
-      content,
-    });
-  };
-
-  // 수정사항 저장
-  const handleEditSave = async () => {
-    if (!editNotice.title || !editNotice.content) {
-      alert(ALERT_TEXT[0]);
-      return;
-    }
-    try {
-      await callPut(`/api/admin/notices/put/${selectedId}`, editNotice);
-      alert(ALERT_TEXT[6]);
-      setEditNotice({
-        title: '',
-        content: '',
-      });
-      setEditingId(null);
-      await handleGetNotice(); // 저장 후 새로 불러오기
-    } catch (error) {
-      console.error(error);
-    }
-  };
-
-  // 공지사항 삭제
-  const handleDelete = async () => {
-    try {
-      await callDelete(`/api/admin/notices/delete/${selectedId}`);
-      alert(ALERT_TEXT[7]);
-      await handleGetNotice(); // 삭제 후 새로 불러오기
-    } catch (error) {
-      console.error(error);
-    }
-  };
   return (
     <div className="flex flex-col items-center gap-4 border-2 p-8">
       {isPost ? (
@@ -159,69 +116,15 @@ export default function NoticeGet() {
               </div>
 
               {selectedId === item.id && (
-                <>
-                  <div className="border-t-[1px] border-gray-4 mt-4 px-2 py-4 flex flex-col gap-4">
-                    {editingId === item.id ? (
-                      <>
-                        <Input
-                          type="default"
-                          className="font-bold border p-2"
-                          textValue={editNotice.title}
-                          onChange={(e) =>
-                            setEditNotice((prev) => ({
-                              ...prev,
-                              title: e.target.value,
-                            }))
-                          }
-                        />
-                        <textarea
-                          className="border-2 py-2 px-4 font-bold min-h-44"
-                          value={editNotice.content}
-                          onChange={(e) =>
-                            setEditNotice((prev) => ({
-                              ...prev,
-                              content: e.target.value,
-                            }))
-                          }
-                        />
-                      </>
-                    ) : (
-                      <>
-                        <p className="border font-bold p-2">{item.title}</p>
-                        <div className="border-2 py-2 px-4">
-                          <p className="font-bold min-h-44">{item.content}</p>
-                        </div>
-                      </>
-                    )}
-                  </div>
-                  <div className="flex gap-4 justify-end">
-                    <Button
-                      className={`${editingId === item.id ? 'bg-gray-2' : 'bg-red-1'} text-white font-bold max-w-fit px-8 py-2 self-end`}
-                      buttonText={
-                        editingId === item.id ? BTN_TEXT[5] : BTN_TEXT[1]
-                      }
-                      type="default"
-                      onClickHandler={
-                        editingId === item.id
-                          ? () => setEditingId(null)
-                          : handleDelete
-                      }
-                    />
-
-                    <Button
-                      className="bg-primary-3 text-white font-bold max-w-fit px-8 py-2 self-end"
-                      buttonText={
-                        editingId === item.id ? BTN_TEXT[6] : BTN_TEXT[3]
-                      }
-                      type="default"
-                      onClickHandler={
-                        editingId === item.id
-                          ? handleEditSave
-                          : () => handleEdit(item.id, item.title, item.content)
-                      }
-                    />
-                  </div>
-                </>
+                <NoticesPut
+                  editingId={editingId}
+                  selectedId={selectedId}
+                  editNotice={editNotice}
+                  setEditNotice={setEditNotice}
+                  setEditingId={setEditingId}
+                  handleGetNotice={handleGetNotice}
+                  item={item}
+                />
               )}
             </div>
           ))}
