+++ src/app/components/admin/notices/NoticePut.tsx
@@ -1,63 +1,52 @@
-import React from 'react';
-import { ALERT_TEXT, BTN_TEXT, FAQ_CATEGORY } from '@/app/constants/admin';
+import { ALERT_TEXT, BTN_TEXT } from '@/app/constants/admin';
 import { callDelete, callPut } from '@/app/utils/callApi';
 import Button from '../../common/Button';
 import Input from '../../common/Input';
 
-export default function FAQPut({
+export default function NoticesPut({
   editingId,
   selectedId,
-  editFAQ,
-  setEditFAQ,
+  editNotice,
   setEditingId,
-  handleGetFAQ,
+  setEditNotice,
+  handleGetNotice,
   item,
-}: FAQPutProps) {
-  const handleEdit = (
-    faq_id: number,
-    category: string,
-    question: string,
-    answer: string,
-  ) => {
-    setEditingId(faq_id);
-    setEditFAQ({
-      category,
-      question,
-      answer,
+}: NoticePutProps) {
+  // 공지사항 수정모드
+  const handleEdit = (notice_id: number, title: string, content: string) => {
+    setEditingId(notice_id);
+    setEditNotice({
+      title,
+      content,
     });
   };
 
+  // 수정사항 저장
   const handleEditSave = async () => {
-    if (
-      !editFAQ.category ||
-      editFAQ.category === '분류 선택' ||
-      !editFAQ.question ||
-      !editFAQ.answer
-    ) {
+    if (!editNotice.title || !editNotice.content) {
       alert(ALERT_TEXT[0]);
       return;
     }
     try {
-      await callPut(`/api/admin/faq/put/${selectedId}`, editFAQ);
+      await callPut(`/api/admin/notices/put/${selectedId}`, editNotice);
       alert(ALERT_TEXT[6]);
-      setEditFAQ({
-        category: '',
-        question: '',
-        answer: '',
+      setEditNotice({
+        title: '',
+        content: '',
       });
       setEditingId(null);
-      await handleGetFAQ(); // 저장 후 새로 불러오기
+      await handleGetNotice(); // 저장 후 새로 불러오기
     } catch (error) {
       console.error(error);
     }
   };
 
-  // FAQ 삭제
+  // 공지사항 삭제
   const handleDelete = async () => {
     try {
-      await callDelete(`/api/admin/faq/delete/${selectedId}`);
+      await callDelete(`/api/admin/notices/delete/${selectedId}`);
       alert(ALERT_TEXT[7]);
-      await handleGetFAQ(); // 삭제 후 새로 불러오기
+      await handleGetNotice(); // 삭제 후 새로 불러오기
     } catch (error) {
       console.error(error);
     }
@@ -67,49 +56,33 @@ export default function FAQPut({
       <div className="border-t-[1px] border-gray-4 mt-4 px-2 py-4 flex flex-col gap-4">
         {editingId === item.id ? (
           <>
-            <select
-              className="font-bold border p-2"
-              value={editFAQ.category}
-              onChange={(e) =>
-                setEditFAQ((prev) => ({
-                  ...prev,
-                  category: e.target.value,
-                }))
-              }
-            >
-              {FAQ_CATEGORY.map((option) => (
-                <option key={option} value={option}>
-                  {option}
-                </option>
-              ))}
-            </select>
             <Input
               type="default"
               className="font-bold border p-2"
-              textValue={editFAQ.question}
+              textValue={editNotice.title}
               onChange={(e) =>
-                setEditFAQ((prev) => ({
+                setEditNotice((prev) => ({
                   ...prev,
-                  question: e.target.value,
+                  title: e.target.value,
                 }))
               }
             />
             <textarea
               className="border-2 py-2 px-4 font-bold min-h-44"
-              value={editFAQ.answer}
+              value={editNotice.content}
               onChange={(e) =>
-                setEditFAQ((prev) => ({
+                setEditNotice((prev) => ({
                   ...prev,
-                  answer: e.target.value,
+                  content: e.target.value,
                 }))
               }
             />
           </>
         ) : (
           <>
-            <p className="border font-bold p-2">{item.question}</p>
+            <p className="border font-bold p-2">{item.title}</p>
             <div className="border-2 py-2 px-4">
-              <p className="font-bold min-h-44">{item.answer}</p>
+              <p className="font-bold min-h-44">{item.content}</p>
             </div>
           </>
         )}
@@ -123,6 +96,7 @@ export default function FAQPut({
             editingId === item.id ? () => setEditingId(null) : handleDelete
           }
         />
+
         <Button
           className="bg-primary-3 text-white font-bold max-w-fit px-8 py-2 self-end"
           buttonText={editingId === item.id ? BTN_TEXT[6] : BTN_TEXT[3]}
@@ -130,8 +104,7 @@ export default function FAQPut({
           onClickHandler={
             editingId === item.id
               ? handleEditSave
-              : () =>
-                  handleEdit(item.id, item.category, item.question, item.answer)
+              : () => handleEdit(item.id, item.title, item.content)
           }
         />
       </div>
